Separate main into its own package
This is easier for vendoring in downstream projects.
